hotfix: float to int as float is incompatible in drawing

+++ mazes/grids/basic_grid.py
@@ -89,7 +89,7 @@ class BasicGrid:
     def to_png(
         self,
         cell_size: int = 15,
-        wall_thickness: float = 1,
+        wall_thickness: int = 1,
         output_name: str = "maze.png",
     ) -> None:
         img_width = cell_size * self.cols
@@ -127,7 +127,7 @@ class BasicGrid:
     def to_gif(
         self,
         cell_size: int = 15,
-        wall_thickness: float = 1,
+        wall_thickness: int = 1,
         duration=5,
         loop=0,
         output_name: str = "maze.gif",
@@ -188,7 +188,7 @@ class BasicGrid:
     def to_vid(
         self,
         cell_size: int = 15,
-        wall_thickness: float = 1,
+        wall_thickness: int = 1,
         duration=5,
         output_name: str = "maze.mp4",
     ) -> None:
