+++ mazes/grids/colored_grid.py
@@ -42,7 +42,7 @@ class ColoredGrid(DistanceGrid):
     def to_png(
         self,
         cell_size: int = 15,
-        wall_thickness: float = 1,
+        wall_thickness: int = 1,
         output_name: str = "maze.png",
         display_distances: bool = False,
         cell_color: tuple[int, int, int] = (255, 0, 0),
@@ -93,7 +93,7 @@ class ColoredGrid(DistanceGrid):
     def to_gif(
         self,
         cell_size: int = 15,
-        wall_thickness: float = 1,
+        wall_thickness: int = 1,
         duration=5,
         loop=0,
         output_name: str = "maze.gif",
@@ -173,7 +173,7 @@ class ColoredGrid(DistanceGrid):
     def to_vid(
         self,
         cell_size: int = 15,
-        wall_thickness: float = 1,
+        wall_thickness: int = 1,
         duration=5,
         output_name: str = "maze.mp4",
         display_distances: bool = False,
